issue powerControlEventHandle error;  issue 256 error;

+++ FUNC_HANDLER/modbus.h
@@ -96,7 +96,7 @@ extern _TKS_FLAGA_type modbusFlag;
 
 void Modbus_UART_Init(void);
 void Modbus_Process_Task(void);
-void getCmd(u8 *index);
+void getCmd(u16 *index);
 void checkChange(void);
 void forcedOutputHnadle(void);
